Refactor | 기본 시스템 파일 내 절대경로 수정 및 코드 구성 변경
- 타입 > 스타일 > 코드 순
