+++ src/pages/_document.tsx
@@ -1,7 +1,7 @@
 import Document, { Head, Html, Main, NextScript } from 'next/document';
 import { ServerStyleSheet } from 'styled-components';
 
-import { Header } from '@/components';
+import Header from '@/components/common/Header';
 import Script from 'next/script';
 import React from 'react';
